:rotating_light: Swiftlint config updated and swiftlint warnings addressed.
