--- GeoTrackKitExample/GeoTrackKitExampleTests/GeoTrackKit/Core/GeoTrackStatisticsTests.swift
@@ -1,14 +1,0 @@
-//
-//  GeoTrackStatisticsTests.swift
-//  GeoTrackKitExampleTests
-//
-//  Created by Eric Internicola on 10/12/18.
-//  Copyright © 2018 Eric Internicola. All rights reserved.
-//
-
-import XCTest
-
-class GeoTrackStatisticsTests: XCTestCase {
-
-
-}
